+++ mopidy/__init__.py
@@ -3,7 +3,7 @@ if not (2, 6) <= sys.version_info < (3,):
     sys.exit(u'Mopidy requires Python >= 2.6, < 3')
 
 def get_version():
-    return u'0.1.0'
+    return u'0.2.0a1'
 
 class MopidyException(Exception):
     def __init__(self, message, *args, **kwargs):
@@ -20,6 +20,9 @@ class MopidyException(Exception):
         self._message = message
 
 class SettingsError(MopidyException):
+    pass
+
+class OptionalDependencyError(MopidyException):
     pass
 
 from mopidy import settings as default_settings_module