+++ mopidy/utils/log.py
@@ -3,27 +3,40 @@ import logging.handlers
 
 from mopidy import settings
 
-def setup_logging(verbosity_level, dump):
+def setup_logging(verbosity_level, save_debug_log):
+    setup_root_logger()
     setup_console_logging(verbosity_level)
-    if dump:
-        setup_dump_logging()
+    if save_debug_log:
+        setup_debug_logging_to_file()
+
+def setup_root_logger():
+    root = logging.getLogger('')
+    root.setLevel(logging.DEBUG)
 
 def setup_console_logging(verbosity_level):
     if verbosity_level == 0:
-        level = logging.WARNING
+        log_level = logging.WARNING
+        log_format = settings.CONSOLE_LOG_FORMAT
     elif verbosity_level == 2:
-        level = logging.DEBUG
+        log_level = logging.DEBUG
+        log_format = settings.DEBUG_LOG_FORMAT
     else:
-        level = logging.INFO
-    logging.basicConfig(format=settings.CONSOLE_LOG_FORMAT, level=level)
-
-def setup_dump_logging():
-    root = logging.getLogger('')
-    root.setLevel(logging.DEBUG)
-    formatter = logging.Formatter(settings.DUMP_LOG_FORMAT)
-    handler = logging.handlers.RotatingFileHandler(
-        settings.DUMP_LOG_FILENAME, maxBytes=102400, backupCount=3)
+        log_level = logging.INFO
+        log_format = settings.CONSOLE_LOG_FORMAT
+    formatter = logging.Formatter(log_format)
+    handler = logging.StreamHandler()
     handler.setFormatter(formatter)
+    handler.setLevel(log_level)
+    root = logging.getLogger('')
+    root.addHandler(handler)
+
+def setup_debug_logging_to_file():
+    formatter = logging.Formatter(settings.DEBUG_LOG_FORMAT)
+    handler = logging.handlers.RotatingFileHandler(
+        settings.DEBUG_LOG_FILENAME, maxBytes=10485760, backupCount=3)
+    handler.setFormatter(formatter)
+    handler.setLevel(logging.DEBUG)
+    root = logging.getLogger('')
     root.addHandler(handler)
 
 def indent(string, places=4, linebreak='\n'):